add edit overview to settings
