+++ frontend_v2/src/app/components/challenge/challengeoverview/challengeoverview.component.html
@@ -10,13 +10,6 @@
             class="challenge-description fs-16 fw-light"
           ></div>
         </div>
-        <div *ngIf="isChallengeHost" class="col-sm-1">
-          <h5>
-            <a class="pointer fs-16" (click)="editChallengeOverview()">
-              <i class="fa fa-pencil" aria-hidden="true"> </i>&nbsp;edit
-            </a>
-          </h5>
-        </div>
       </div>
     </div>
   </div>
