+++ frontend_v2/src/app/components/challenge/challengeoverview/challengeoverview.component.ts
@@ -5,7 +5,6 @@ import { NGXLogger } from 'ngx-logger';
 // import service
 import { GlobalService } from '../../../services/global.service';
 import { ChallengeService } from '../../../services/challenge.service';
-import { ApiService } from '../../../services/api.service';
 import { EndpointsService } from '../../../services/endpoints.service';
 
 /**
@@ -28,11 +27,6 @@ export class ChallengeoverviewComponent implements OnInit {
   isChallengeHost = false;
 
   /**
-   * To call the API inside modal for editing the challenge description
-   */
-  apiCall: any;
-
-  /**
    * Constructor.
    * @param document  Window document Injection.
    * @param challengeService  ChallengeService Injection.
@@ -41,7 +35,6 @@ export class ChallengeoverviewComponent implements OnInit {
     private challengeService: ChallengeService,
     @Inject(DOCUMENT) private document: Document,
     private globalService: GlobalService,
-    private apiService: ApiService,
     private endpointsService: EndpointsService,
     private logger: NGXLogger
   ) {}
@@ -56,37 +49,5 @@ export class ChallengeoverviewComponent implements OnInit {
     this.challengeService.isChallengeHost.subscribe((status) => {
       this.isChallengeHost = status;
     });
-  }
-
-  editChallengeOverview() {
-    const SELF = this;
-
-    SELF.apiCall = (params) => {
-      const BODY = JSON.stringify(params);
-      SELF.apiService
-        .patchUrl(SELF.endpointsService.editChallengeDetailsURL(SELF.challenge.creator.id, SELF.challenge.id), BODY)
-        .subscribe(
-          (data) => {
-            SELF.challenge.description = data.description;
-            SELF.globalService.showToast('success', 'The description is successfully updated!', 5);
-          },
-          (err) => {
-            SELF.globalService.handleApiError(err, true);
-            SELF.globalService.showToast('error', err);
-          },
-          () => this.logger.info('EDIT-CHALLENGE-DESCRIPTION-FINISHED')
-        );
-    };
-
-    const PARAMS = {
-      title: 'Edit Challenge Description',
-      label: 'description',
-      isEditorRequired: true,
-      editorContent: this.challenge.description,
-      confirm: 'Submit',
-      deny: 'Cancel',
-      confirmCallback: SELF.apiCall,
-    };
-    SELF.globalService.showModal(PARAMS);
   }
 }
