+++ frontend_v2/src/app/components/challenge/challengesettings/challengesettings.component.html
@@ -69,6 +69,17 @@
           (click)="stopParticipation($event)"
           ><strong class="text-light-black fs-15">Close Participation</strong>
         </mat-checkbox>
+        <div class="ev-action">
+          <div *ngIf="isChallengeHost" class="col-sm-1.5 col-xs-1.5 ev-upload-action">
+            <span class="fs-16 fw-light">
+              <strong class="text-light-black">Overview: </strong>
+            </span>
+            <a class="pointer fs-16" (click)="editChallengeOverview()">
+              &nbsp;&nbsp;
+              <i class="fa fa-pencil" aria-hidden="true"> </i>&nbsp;Edit
+            </a>
+          </div>
+        </div>
       </div>
     </div>
